move HandlerResultFillerRegistry to package filler
